pyproject.toml
[project]
name = "monoco-toolkit"
version = "0.4.0"
description = "The Headless Operating System for Agentic Engineering"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = [
    "monoco",
    "agent-native",
    "task-management",
    "kanban",
    "ai-agents",
    "workflow",
    "cli",
]
authors = [
    { name = "Monoco Team", email = "dev@monoco.io" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Office/Business :: Groupware",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "prompt-toolkit>=3.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "sse-starlette>=1.6.0",
    "watchdog>=6.0.0",
    "httpx>=0.28.1",
]

[project.urls]
Homepage = "https://monoco.io"
Repository = "https://github.com/IndenScale/Monoco"
Documentation = "https://monoco.io/docs"
Issues = "https://github.com/IndenScale/Monoco/issues"

[project.scripts]
monoco = "monoco.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/monoco"]

[dependency-groups]
dev = [
    "pytest-mock>=3.14.0",
    "pre-commit>=3.5.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=4.1.0",
]

[tool.ruff.lint]
ignore = ["E402", "E722", "F811", "F841", "E741"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["tests/legacy", "Archives", ".git", ".venv"]
